mini-server-prod/kgroot-services/rca-api/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Neo4j
neo4j==5.26.0

# OpenAI (GPT-5)
openai==1.55.0

# Embeddings and vector search
sentence-transformers==3.3.1
numpy==2.1.3
faiss-cpu==1.9.0

# Slack integration
slack-sdk==3.33.4

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.17
httpx==0.27.2
redis==5.2.0
tenacity==9.0.0

# Monitoring
prometheus-client==0.21.0